refactor(evidence): move models and rules into domain
