+++ src/agent_memory_orchestrator/application/services/session/detail.py
@@ -6,14 +6,14 @@ from pathlib import Path
 from typing import Any
 
 from ....core.config import Settings
-from ....evidence.drain import _read_jsonl_from
-from ....evidence.triggers import TriggerDecision
-from ....evidence.triggers import is_session_start
-from ....evidence.triggers import record_session_id
-from ....evidence.triggers import session_boundary_trigger
-from ....evidence.window import clean_evidence_window
-from ....infrastructure.sqlite.production_job_store import ProductionSessionJobStore
+from ....domain.evidence import TriggerDecision
+from ....domain.evidence import clean_evidence_window
+from ....domain.evidence import is_session_start
+from ....domain.evidence import record_session_id
+from ....domain.evidence import session_boundary_trigger
 from ....domain.retrieval.text import clip_text as _clip
+from ....evidence.drain import _read_jsonl_from
+from ....infrastructure.sqlite.production_job_store import ProductionSessionJobStore
 
 
 def _load_evidence_records(
