--- src/agent_memory_orchestrator/domain/evidence/boundaries.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from ...evidence import EvidenceDrain
-from ...evidence import RawEvidenceStore
-from ...evidence import clean_evidence_window
-
-__all__ = ["EvidenceDrain", "RawEvidenceStore", "clean_evidence_window"]
